UGENE-715 is fixed
git-svn-id: https://local.ugene.unipro.ru/svn/ugene/trunk@1755 7ca3a678-163c-e011-b84f-001e8c0267ff

+++ src/corelibs/U2Core/src/datatype/U2Region.cpp
@@ -80,7 +80,7 @@ void U2Region::reverse(QVector<U2Region>& regions) {
     }
 }
 
-void U2Region::shift(int offset, QVector<U2Region>& regions) {
+void U2Region::shift(qint64 offset, QVector<U2Region>& regions) {
     QVector<U2Region> res;
     for (int i = 0, n = regions.size(); i < n; i++) {
         U2Region& r = regions[i];
